Add locators.properties file
locators.properties has been added to the project to remove the locators
hard coding from the test and manage them externally.

BastTest class has been updated to read in the file and create a
locators object which can be used in the test.

The LoginTest has been updated using the locators from an external source.