Add ReadXlsData
This commit adds the ReadXlsData class to the project.

This allows us to read in an excel file and gather the date within
each row and column to use for our test data.  Further externalizing
the data of the project.

The LoginTestData.xlsx file has been added which contains a list
of usernames and passwords to be used in the LoginTest